fix(ui): dedupe cluster-scoped list hook to clear jscpd clone
The Headlamp-compat list surfaces added in #5385 inlined the same
cluster-scoped fetch-on-change React effect in two places —
makeResourceClass.useList (k8s.ts) and makeK8sShim.useResourceList
(pluginLib.ts) — differing only in the item type, the fetch call, and the
extra effect deps. jscpd (MegaLinter COPYPASTE, threshold 0%) flagged the
two blocks as clones, failing the `🧹 Lint - mega-linter` check on every
open PR (MegaLinter full-scans the repo on each PR's merge ref).

Extract the shared effect into a generic useClusterScopedList<T>(getCluster,
fetchList, deps) hook and have both call sites delegate to it. Behaviour is
preserved: same [items, error] return, same cluster-keyed re-fetch, same
extra deps (kind/namespace for useResourceList). Drops the now-unused React
import from k8s.ts.

Validated: web/ui `tsc --noEmit` clean; jscpd reports 0 clones (was 2).
